+++ lang/da/pagination.php
@@ -7,5 +7,6 @@ return [
     'of' => 'af',
     'many' => 'mange',
     'show' => 'vis',
+    'results' => 'resultater',
 
 ];